requirements.txt
flask==2.0.1
python-dotenv==0.19.1
requests==2.26.0
pandas==1.3.3
numpy==1.21.2
python-dateutil==2.8.2
gunicorn==20.1.0 